+++ mpc_energy/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.0
 * Added net grid interaction kWh sensor to aid automations.
 * Added EV charging
+* Removed support for RBC controller due to lack of development.
 
 ## 0.6.0
 * Adding flow power as retailer option